+++ src/App.jsx
@@ -2,17 +2,46 @@ import './App.css';
 import MemoIndex from './components/MemoIndex.jsx';
 import MemoEditor from './components/MemoEditor.jsx';
 import useMemos from './hooks/useMemos.js';
+import { useState } from 'react';
 
 export default function App() {
-  const {
-    memos,
-    editingMemo,
-    setEditingMemo,
-    handleMemoClick,
-    handleAddButtonClick,
-    handleUpdateButtonClick,
-    handleDeleteButtonClick,
-  } = useMemos();
+  const [editingMemo, setEditingMemo] = useState();
+  const { memos, addMemo, updateMemo, deleteMemo } = useMemos();
+
+  function handleMemoClick(e, memo) {
+    e.preventDefault();
+    setEditingMemo(memo);
+  }
+
+  function handleAddButtonClick(e) {
+    e.preventDefault();
+    const maxId = Math.max(-1, ...memos.map((m) => m.id)) + 1;
+    const nextMemo = {
+      id: maxId,
+      content: '新規メモ',
+    };
+    addMemo(nextMemo);
+    setEditingMemo(nextMemo);
+  }
+
+  function handleUpdateButtonClick(memo) {
+    const content = memo.content.trim();
+    if (content === '') {
+      alert('空白では入力できません');
+      return;
+    }
+    const updatedMemo = {
+      id: memo.id,
+      content,
+    };
+    updateMemo(updatedMemo);
+    setEditingMemo(updatedMemo);
+  }
+
+  function handleDeleteButtonClick(memo) {
+    deleteMemo(memo);
+    setEditingMemo();
+  }
 
   return (
     <>
