+++ src/hooks/useMemos.js
@@ -1,4 +1,4 @@
-import { useState, useEffect, useReducer } from 'react';
+import { useEffect, useReducer } from 'react';
 import memosReducer from '../reducers/memosReducer.js';
 import {
   setItemsToLocalStorage,
@@ -6,7 +6,6 @@ import {
 } from '../utils/localStorage.js';
 
 export default function useMemos() {
-  const [editingMemo, setEditingMemo] = useState();
   const [memos, dispatch] = useReducer(
     memosReducer,
     getItemsFromLocalStorage('memos'),
@@ -16,57 +15,31 @@ export default function useMemos() {
     setItemsToLocalStorage('memos', memos);
   }, [memos]);
 
-  function handleMemoClick(e, memo) {
-    e.preventDefault();
-    setEditingMemo(memo);
-  }
-
-  function handleAddButtonClick(e) {
-    const maxId = Math.max(-1, ...memos.map((m) => m.id)) + 1;
-    e.preventDefault();
-    const nextMemo = {
-      id: maxId,
-      content: '新規メモ',
-    };
+  function addMemo(memo) {
     dispatch({
       type: 'added',
-      ...nextMemo,
+      ...memo,
     });
-    setEditingMemo(nextMemo);
   }
 
-  function handleUpdateButtonClick(memo) {
-    const content = memo.content.trim();
-    if (content === '') {
-      alert('空白では入力できません');
-      return;
-    }
-    const updatedMemo = {
-      id: memo.id,
-      content,
-    };
+  function updateMemo(memo) {
     dispatch({
       type: 'updated',
-      memo: updatedMemo,
+      memo: memo,
     });
-    setEditingMemo(updatedMemo);
   }
 
-  function handleDeleteButtonClick(memo) {
+  function deleteMemo(memo) {
     dispatch({
       type: 'deleted',
       id: memo.id,
     });
-    setEditingMemo();
   }
 
   return {
     memos,
-    editingMemo,
-    setEditingMemo,
-    handleMemoClick,
-    handleAddButtonClick,
-    handleUpdateButtonClick,
-    handleDeleteButtonClick,
+    addMemo,
+    updateMemo,
+    deleteMemo,
   };
 }
